requirements.txt
# Core dependencies
requests>=2.25.1
python-dotenv>=0.19.0
tenacity>=8.0.1
numpy>=1.20.0
pandas>=1.3.0

# Optional time-series database
# influxdb-client>=1.24.0
# timescaledb>=0.3.0

# API integration
websocket-client>=1.2.1
aiohttp>=3.8.1

# Testing and development
pytest>=6.2.5
pytest-mock>=3.6.1

# Monitoring
# prometheus-client>=0.11.0

# Queue system (for production deployment)
pika>=1.2.0
kafka-python>=2.0.2
aio-pika>=9.0.0

# Security and encryption
cryptography>=35.0.0

# Utilities
tabulate>=0.8.9

# Additional dependencies
pyyaml>=6.0
schedule>=1.1.0
sqlalchemy>=1.4.0
ccxt>=1.60.0
websockets>=10.0
colorama>=0.4.4
tqdm>=4.62.0
plotly>=5.3.0
responses>=0.13.0
asyncio>=3.4.3
aiosqlite>=0.17.0
matplotlib>=3.5.0
ujson>=5.0.0
pytest-asyncio>=0.18.0
coverage>=6.0.0
pre-commit>=2.17.0
black>=22.1.0
mypy>=0.931
isort>=5.10.0
flake8>=4.0.0 